Fix all relevant GCC warnings (LA.UM.6.4.r1-06900-8x98.0)
This is the first part in the GCC warnings series, fixing all
warnings that come from CAF and Linux upstream drivers.

This was done with the following toolchain version:
aarch64-linux-gnu-gcc (Linaro GCC 7.2-2017.11) 7.2.1 20171011

If a warning was a false positive, it was disabled in the folder's
Makefile. I believe all of these are correct but I am 100% open to
discussion (I am not infallible).