oneplus5: Skip sync on suspend
As per Motorola:

"On embedded devices with built-in batteries, it is not so
important to sync the file systems before suspend.  The chance
of losing power during suspend are no greater than they are
when the system is awake.  The sync operations can greatly
increase suspend latency when the system has accrued many dirty
pages and/or the target storage devices are not particularly
fast."

Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>