HACK: sched: core: Disable constant-logical-operand
When CONFIG_SCHED_DEBUG is disabled, this warning gets triggered
because of the SCHED_FEAT macro. Commit 765cc3a4b224 ("sched/core:
Optimize sched_feat() for !CONFIG_SCHED_DEBUG builds") in mainline
seems like it tries to address this but it doesn't work here. Just
disable the warning for now.